Use f64 clocks for long-running processes; unwrap the Pulse tick rollover; add PBEngineInfo.total_runtime_f64
The headless authoritative server runs for weeks, and the engine kept every
clock as f32 seconds since process start. At a 30 Hz tick the f32 resolution
reaches the tick period after ~3 days and exceeds it after ~6, at which point
the scene scheduler's `last_sent < elapsed` gate reads false on repeated
frames and its priority divides by zero, and the foreign-avatar interpolation
windows degrade the same way.

- Read `Time::elapsed_secs_f64()` everywhere and widen the fields those
  reads land in (scene dispatch/runtime stamps, foreign-avatar target and
  receipt times, jump time, emote timestamp, toasts, chat history, pointer
  lock, gizmo cache, ...). Intervals stay f32; deltas are cast at the point
  of use so the Vec3 math is unchanged. Time-of-day (seconds since midnight,
  bounded) is deliberately left f32.
- Carry `SceneElapsedTime` as f64 across the scene IPC (native and wasm).
  The message-pack channel is positional, so the two native binaries must
  ship together.
- Pulse `server_tick` is u32 milliseconds since the Pulse server started and
  rolls over after ~49.7 days; the receiver's strictly-newer gate would then
  drop every update. The decoder now unwraps it with a wrapping half-range
  test, adding 2^32 ms per rollover. Numerically identical to the old value
  until the first rollover.
- Add `double total_runtime_f64 = 5` to PBEngineInfo (protocol PR alongside).
  The existing float field can't be changed in place: the SDK's generated
  decoder guards the wire type and would skip it, so scenes would read 0.

+++ crates/common/src/structs.rs
@@ -1236,7 +1236,7 @@ pub struct EngineMovementControl {
     /// is received. Set by `movePlayerTo` when it imposes a facing, so the imposed
     /// orientation survives until the controller scene reads the new transform and
     /// echoes it back, rather than being clobbered by an in-flight stale tick.
-    pub accept_movement_after: f32,
+    pub accept_movement_after: f64,
 }
 
 #[derive(Default, Clone, Copy, PartialEq, Eq)]
@@ -1264,7 +1264,7 @@ pub enum MoveKind {
 pub struct AvatarDynamicState {
     pub velocity: Vec3,
     pub ground_height: f32,
-    pub jump_time: f32,
+    pub jump_time: f64,
     pub move_kind: MoveKind,
 }
